pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Rusin Vladislav"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.1.0"
django-filter = "^22.1"
pydantic = "^1.10.7"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.pytest.ini_options]
#DJANGO_SETTING_MODULE = 'todolist.settings'
#python_files = [
#    'tests.py',
#    'test_*.py',
#    '*_tests.py',
#]
